OCLOMRS-929: Obscure api token on user profile page
This change uses a simple blur effect to obscure the user's API token on their profile page. The token is visible on hover and can be copied to clipboard just as before. This simple approach can help protect user API tokens from being accidentally compromised until we come up with a better approach.